requirements.txt
PyYAML~=6.0
setuptools~=63.2.0